# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SaaS backend with Keycloak-issued JWT auth.\nAll protected endpoints require a Bearer token obtained from Keycloak.",
        "title": "Lightweight SaaS Backend API",
        "contact": {},
        "version": "0.4.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a page of users from the Keycloak realm. Requires the\ncaller's token to carry the realm `admin` role. Pagination\nis offset-based; `max` is clamped to [1, 100] server-side.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List users (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "substring match on username/email/firstName/lastName",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset (default 0)",
                        "name": "first",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size (default 20, max 100)",
                        "name": "max",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/identity.ListUsersResponse"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "token lacks admin role",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "upstream identity provider unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "identity management not configured",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "identity.ListUsersResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "length of users array, NOT total",
                    "type": "integer"
                },
                "first": {
                    "type": "integer"
                },
                "max": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/identity.UserResponse"
                    }
                }
            }
        },
...
        "identity.UserResponse": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "OPERATOR authentication. Type \"Bearer\" followed by a Keycloak-issued access token. The token must carry the realm `admin` role, which is re-checked against the provider on every request.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ProjectKeyAuth": {
            "description": "PROJECT authentication (machine-to-machine). Type \"Bearer\" followed by a project credential, e.g. `Bearer lw_sk_\u003clookup\u003e_\u003csecret\u003e`. The credential is bound to exactly one workspace and carries an explicit scope set; the scope each operation requires is stated in its description. Control-plane operations (workspaces, connections, projects) are operator-only and answer `operator_only` to any credential. Swagger 2.0 cannot express per-operation scopes for an apiKey scheme, so they are documented in prose rather than declared — see docs/PROJECTS.md.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}